Debug Graph(uncomment in Game.cpp) Actor usable abilities - yet to implement any Fixed some rounding issues

+++ include/engine/utils/variable/datatype/Macros.h
@@ -2,10 +2,10 @@
 
 typedef unsigned char Uint8;
 typedef unsigned short Uint16;
-typedef unsigned int Uint32;
-typedef unsigned long Uint64;
+typedef unsigned long Uint32;
+typedef unsigned long long Uint64;
 
 typedef signed char Sint8;
 typedef signed short Sint16;
-typedef signed int Sint32;
-typedef signed long Sint64;
+typedef signed long Sint32;
+typedef signed long long Sint64;
